refactor: move StructBuilder from builder to parser
