+++ mapillary_tools/geotag/simple_mp4_builder.py
@@ -12,176 +12,16 @@ import construct as C
 
 from . import io_utils, simple_mp4_parser as parser
 from .simple_mp4_parser import (
-    ChunkLargeOffsetBox,
-    ChunkOffsetBox,
-    DataEntryUrlBox,
-    DataEntryUrnBox,
-    DataReferenceBox,
-    EditBox,
-    HandlerReferenceBox,
-    MediaHeaderBox,
-    MovieHeaderBox,
+    BoxHeader32,
+    BoxHeader64,
+    FullBoxStruct32,
+    QuickBoxStruct32,
+    QuickBoxStruct64,
     RawSample,
-    SampleDescriptionBox,
-    SampleSizeBox,
-    SampleToChunkBox,
-    SyncSampleBox,
-    TimeToSampleBox,
-    TrackHeaderBox,
 )
 
 UINT32_MAX = 2**32 - 1
 UINT64_MAX = 2**64 - 1
-
-
-BoxHeader0 = C.Struct(
-    "size32" / C.Const(0, C.Int32ub),
-    "type" / C.Bytes(4),
-)
-
-BoxHeader32 = C.Struct(
-    "size" / C.Int32ub,
-    "type" / C.Bytes(4),
-)
-
-BoxHeader64 = C.Struct(
-    "size32" / C.Const(1, C.Int32ub),
-    "type" / C.Bytes(4),
-    "size" / C.Int64ub,
-)
-
-
-class Box64StructBuilder:
-    """
-    Build a box struct that **parses** MP4 boxes with both 32-bit and 64-bit sizes.
-
-    NOTE: Do not build data with this struct. For building, use Box32StructBuilder instead.
-    """
-
-    _box: C.Struct
-
-    def __init__(
-        self, switch_map: T.Dict[bytes, C.Struct], lazy_box_types: T.Sequence[bytes]
-    ) -> None:
-        self._box = None
-        self._box_type_switch_map = {**switch_map}
-        lazy_box = C.LazyBound(lambda: C.GreedyRange(self.Box))
-        for box_type in lazy_box_types:
-            self._box_type_switch_map[box_type] = lazy_box
-        self._switch = C.Switch(
-            C.this.type,
-            self._box_type_switch_map,
-            C.GreedyBytes,
-        )
-
-    @property
-    def Box(self) -> C.Struct:
-        if self._box is None:
-            BoxData32 = C.Struct(
-                "data"
-                / C.FixedSized(
-                    C.this.size - 8,
-                    self._switch,
-                )
-            )
-
-            BoxData64 = C.Struct(
-                "data"
-                / C.FixedSized(
-                    C.this.size - 16,
-                    self._switch,
-                )
-            )
-
-            BoxData0 = C.Struct(
-                "data" / self._switch,
-            )
-
-            self._box = C.Select(
-                BoxHeader32 + BoxData32, BoxHeader64 + BoxData64, BoxHeader0 + BoxData0
-            )
-
-        return self._box
-
-    @property
-    def BoxList(self) -> C.Struct:
-        return C.GreedyRange(self.Box)
-
-
-class Box32StructBuilder(Box64StructBuilder):
-    """
-    Build a box struct that parses or builds MP4 boxes with 32-bit size only.
-
-    NOTE: The struct does not handle extended size correctly.
-    To parse boxes with extended size, use Box64StructBuilder instead.
-    """
-
-    @property
-    def Box(self) -> C.Struct:
-        if self._box is None:
-            self._box = C.Prefixed(
-                C.Int32ub,
-                C.Struct("type" / C.Bytes(4), "data" / self._switch),
-                includelength=True,
-            )
-
-        return self._box
-
-
-_full_switch_map = {
-    b"tkhd": TrackHeaderBox,
-    b"mdhd": MediaHeaderBox,
-    b"stsc": SampleToChunkBox,
-    b"stts": TimeToSampleBox,
-    b"co64": ChunkLargeOffsetBox,
-    b"stco": ChunkOffsetBox,
-    b"stsd": SampleDescriptionBox,
-    b"stsz": SampleSizeBox,
-    b"stss": SyncSampleBox,
-    b"hdlr": HandlerReferenceBox,
-    b"dref": DataReferenceBox,
-    b"urn ": DataEntryUrnBox,
-    b"url ": DataEntryUrlBox,
-    b"mvhd": MovieHeaderBox,
-    b"elst": EditBox,
-}
-_full_lazy_box_types = [
-    b"moov",
-    b"trak",
-    b"edts",
-    b"mdia",
-    b"minf",
-    b"stbl",
-    b"mvex",
-    b"moof",
-    b"traf",
-    b"mfra",
-    b"dinf",
-    b"edts",
-]
-
-FullBoxStruct32 = Box32StructBuilder(_full_switch_map, _full_lazy_box_types)
-FullBoxStruct64 = Box64StructBuilder(_full_switch_map, _full_lazy_box_types)
-
-_quick_switch_map = {
-    b"tkhd": TrackHeaderBox,
-    b"mdhd": MediaHeaderBox,
-    b"hdlr": HandlerReferenceBox,
-    b"mvhd": MovieHeaderBox,
-    b"elst": EditBox,
-}
-
-_quick_lazy_box_types = [
-    b"moov",
-    b"trak",
-    b"mdia",
-    b"minf",
-    b"mvex",
-    b"edts",
-]
-
-QuickBoxStruct32 = Box32StructBuilder(_quick_switch_map, _quick_lazy_box_types)
-QuickBoxStruct64 = Box64StructBuilder(_quick_switch_map, _quick_lazy_box_types)
 
 
 class BoxDict(TypedDict, total=False):
